requirements.txt.txt
Flask==2.3.2
Flask-CORS==3.1.1
openai==0.27.4
python-dotenv==1.0.0